+++ tests/unit/Framework/Application/Command/RegisterApplicationCest.php
@@ -1,34 +1,19 @@
 <?php
 
+namespace Test\Framework\Application\Command;
+
 use Pckg\Framework\Test\ContextDiff;
 use Pckg\Framework\Test\ListenForEvents;
 use Pckg\Framework\Test\MockConfig;
-use Pckg\Framework\Test\MockFramework;
 use Pckg\Framework\Test\MockInContext;
 
-class FrameworkApplicationCommandRegisterApplicationCest
+class RegisterApplicationCest extends \Pckg\Framework\Test\Codeception\Cest
 {
 
-    use MockFramework;
     use ContextDiff;
     use ListenForEvents;
     use MockConfig;
     use MockInContext;
-
-    protected UnitTester $unitTester;
-
-    public function _before(UnitTester $I)
-    {
-        if (!defined('__ROOT__')) {
-            define('__ROOT__', realpath(__DIR__ . '/../..') . '/');
-        }
-        $this->unitTester = $I;
-        $this->mockFramework();
-    }
-
-    protected function _after()
-    {
-    }
 
     protected function testRegistersWebsiteApplication()
     {
