+++ tests/unit/Framework/Application/WebsiteCest.php
@@ -1,0 +1,30 @@
+<?php
+
+namespace Test\Framework\Application;
+
+use Pckg\Framework\Test\Codeception\Cest;
+use Pckg\Framework\Test\ContextDiff;
+use Pckg\Framework\Test\ListenForEvents;
+use Pckg\Framework\Test\MockConfig;
+use Pckg\Framework\Test\MockInContext;
+
+class WebsiteCest extends Cest
+{
+
+    use ContextDiff;
+    use ListenForEvents;
+    use MockConfig;
+    use MockInContext;
+
+    const EVENT_MIDDLEWARE = self::class . '.middleware';
+    const EVENT_AFTERWARE = self::class . '.afterware';
+
+    public function testWebsite()
+    {
+        $provider = new \Pckg\Framework\Provider();
+        $website = new \Pckg\Framework\Application\Queue($provider);
+
+        $this->unitTester->assertNotEmpty($website->inits());
+        $this->unitTester->assertNotEmpty($website->runs());
+    }
+}
